// src/models/HealthModal.ts
export interface HealthModal {
  partner: number | null;
  partner_details: any;
  serial_no: number;
  child_info: any;
  child_details: any;
  vitamin_A: boolean;
  deworming: boolean;
  ifa: boolean;
  vitamin_A_details: any;
  deworming_details: any;
  ifa_details: any;
  health_camp_date: string;
}