// src/models/ProgramModal.ts
export interface ProgramModal {
  type: string;
  partner: number;
  date: string;
  visiting_team_size: string;
  liaison: any;
  children_participated: string;
  avg_attendance: string;
  enrollers_count: string;
  dropouts_count: number;
  sick_count: string;
  illness: string;
  activity_sheet_no: string;
  is_activity_completed: string;
  is_poshan_calendar_maintained: string;
  food_received_timestamp: string;
  meals_carry_forward: string;
  meals_received: string;
  is_food_safely_stored: string;
  is_breakfast_served_daily: string;
  breakfast_served_at: string;
  additional_info: string;
  teacher_or_social_worker_feedback: string;
  parents_feedback: string;
  children_feedback: string;
  volunteer_details: any;
  visit_duration: string;
}